write some properties tests

+++ src/main/java/org/qw3rtrun/aub/engine/vectmath/Near.java
@@ -1,7 +1,7 @@
 package org.qw3rtrun.aub.engine.vectmath;
 
 public interface Near<T> {
-    boolean isNearTo(T o, float epsilon);
+    boolean isNearTo(T o, double epsilon);
 
     float bound();
 }
